+++ src/main/java/framework/Waiter.java
@@ -1,6 +1,5 @@
-package framework.factory;
+package framework;
 
-import framework.BaseEntity;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.ui.ExpectedConditions;
