+++ src/test/java/TUTTest.java
@@ -1,6 +1,5 @@
 import framework.BaseEntity;
-import framework.factory.BrowserFactory;
-import framework.factory.users.User;
+import framework.dataFactory.User;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 import pages.LoginPage;
@@ -10,12 +9,11 @@ import java.util.List;
 
 
 public class TUTTest extends BaseEntity {
-    private final BrowserFactory factory = BrowserFactory.getInstance();
 
     @DataProvider(name = "getUsers")
     public Object[][] getUsers() {
         return new Object[][]{{
-                BrowserFactory.getUserList()
+//                DataStash.getUserList()
         }
         };
     }
